CONTRACTS: Add the `dfcc_swap_and_wrapt class`.
Given a `function_id`, a `contract_id`, a contract `mode`,
and a `dfcc_contract_handlert` for the contract,
this class swaps the body of the function to a function
with a mangled name, instruments the mangled function
for dynamic frame condition checking using `dfcc_instrument`.
The body of the original function is then replaced with
instructions generated by the contract handler for the
required contract  mode.
To handle recursive functions in contract checking mode,
the body of the wrapper function is generated in such a way
the first invocation results in checking the contract against the
mangled function and recursive invocations of the wrapper
behave like the contract abstraction (modelling that the contract holds).